Add camada ViewModel. Add estrutura base para ViewModels Add tela Categoria

+++ Domain/Entity/DetalheInvestimento.cs
@@ -12,9 +12,9 @@ namespace Domain.Entity
         public long PK_DetalheInvestimento { get; set; }
         
         [Obrigatorio, Relacionamento("Transacao", "PK_Transacao")]
-        public int FK_Transacao { get; set; }
+        public long FK_Transacao { get; set; }
         [Obrigatorio, Relacionamento("AtivoFinanceiro", "PK_AtivoFinanceiro")]
-        public int FK_AtivoFinanceiro { get; set; }
+        public long FK_AtivoFinanceiro { get; set; }
 
         [Obrigatorio, TamanhoDecimal(18,2)]
         public decimal QuantidadeCotas { get; set; }
